+++ www/src/components/banner.js
@@ -1,16 +1,15 @@
 import React from "react"
-import PropTypes from "prop-types"
 import styled from "@emotion/styled"
+import { OutboundLink } from "gatsby-plugin-google-analytics"
 
-import presets, { colors } from "../utils/presets"
-import { rhythm, scale, options } from "../utils/typography"
+import presets, { colors, space } from "../utils/presets"
+import { rhythm, options } from "../utils/typography"
 
-const horizontalPadding = rhythm(1 / 2)
-const backgroundColor = props =>
-  props.background ? props.background : colors.gatsby
+const horizontalPadding = rhythm(space[6])
+const backgroundColor = colors.gatsby
 
 const BannerContainer = styled(`div`)`
-  background-color: ${props => backgroundColor(props)};
+  background-color: ${backgroundColor};
   height: ${presets.bannerHeight};
   position: fixed;
   width: 100%;
@@ -24,30 +23,39 @@ const InnerContainer = styled(`div`)`
   display: flex;
   height: ${presets.bannerHeight};
   overflow-x: auto;
-  mask-image: ${`linear-gradient(to right, transparent, ${props =>
-    backgroundColor(props)} ${horizontalPadding}, ${props =>
-    backgroundColor(props)} 96%, transparent)`};
+  mask-image: ${`linear-gradient(to right, transparent, ${backgroundColor} ${horizontalPadding}, ${backgroundColor} 96%, transparent)`};
 `
 
 const Content = styled(`div`)`
   color: ${colors.ui.bright};
   font-family: ${options.headerFontFamily.join(`,`)};
-  font-size: ${scale(-1 / 5).fontSize};
   padding-left: ${horizontalPadding};
   padding-right: ${horizontalPadding};
   -webkit-font-smoothing: antialiased;
   white-space: nowrap;
+
+  a {
+    color: #fff;
+    border-bottom: 1px solid #fff;
+  }
+
+  a:hover {
+    color: #fff;
+    border-bottom-color: #ffffffa0;
+  }
 `
 
-const Banner = ({ children, background }) => (
-  <BannerContainer background={background} className="banner">
-    <InnerContainer>{children && <Content>{children}</Content>}</InnerContainer>
+const Banner = () => (
+  <BannerContainer className="banner">
+    <InnerContainer>
+      <Content>
+        <OutboundLink href="https://www.gatsbyjs.com/behind-the-scenes/">
+          Watch now
+        </OutboundLink>
+        {`: “Behind the Scenes: What makes Gatsby Great”.`}
+      </Content>
+    </InnerContainer>
   </BannerContainer>
 )
-
-Banner.propTypes = {
-  children: PropTypes.node.isRequired,
-  background: PropTypes.any,
-}
 
 export default Banner