+++ www/src/components/layout.js
@@ -1,9 +1,10 @@
 import React from "react"
+import { navigate, PageRenderer } from "gatsby"
+import mousetrap from "mousetrap"
 import Modal from "react-modal"
 import { SkipNavLink } from "@reach/skip-nav"
-import { OutboundLink } from "gatsby-plugin-google-analytics"
 import MdClose from "react-icons/lib/md/close"
-import { navigate, PageRenderer } from "gatsby"
+
 import presets, { colors } from "../utils/presets"
 import Banner from "../components/banner"
 import Navigation from "../components/navigation"
@@ -11,16 +12,12 @@ import MobileNavigation from "../components/navigation-mobile"
 import PageWithSidebar from "../components/page-with-sidebar"
 import SiteMetadata from "../components/site-metadata"
 
-import mousetrap from "mousetrap"
-
 // Import Futura PT typeface
 import "../fonts/Webfonts/futurapt_book_macroman/stylesheet.css"
 import "../fonts/Webfonts/futurapt_bookitalic_macroman/stylesheet.css"
 import "../fonts/Webfonts/futurapt_demi_macroman/stylesheet.css"
 import "../fonts/Webfonts/futurapt_demiitalic_macroman/stylesheet.css"
-
-// Other fonts
-import "typeface-spectral"
+import { rhythm } from "../utils/typography"
 
 let windowWidth
 
@@ -57,8 +54,6 @@ class DefaultLayout extends React.Component {
   }
 
   render() {
-    const isHomepage = this.props.location.pathname === `/`
-
     // SEE: template-docs-markdown for why this.props.isSidebarDisabled is here
     const isSidebarDisabled =
       this.props.isSidebarDisabled || !this.props.itemList
@@ -72,7 +67,7 @@ class DefaultLayout extends React.Component {
 
     if (isModal && window.innerWidth > 750) {
       return (
-        <React.Fragment>
+        <>
           <PageRenderer
             location={{ pathname: this.props.modalBackgroundPath }}
           />
@@ -109,8 +104,8 @@ class DefaultLayout extends React.Component {
           >
             <div
               css={{
-                backgroundColor: `#ffffff`,
-                borderRadius: presets.radius,
+                backgroundColor: `#fff`,
+                borderRadius: presets.radii[1],
                 boxShadow: `0 0 90px -24px ${colors.gatsby}`,
                 position: `relative`,
               }}
@@ -120,8 +115,8 @@ class DefaultLayout extends React.Component {
                 css={{
                   background: colors.ui.bright,
                   border: 0,
-                  borderBottomLeftRadius: presets.radius,
-                  borderTopRightRadius: presets.radius,
+                  borderBottomLeftRadius: presets.radii[1],
+                  borderTopRightRadius: presets.radii[1],
                   color: colors.gatsby,
                   cursor: `pointer`,
                   position: `absolute`,
@@ -142,37 +137,30 @@ class DefaultLayout extends React.Component {
               {this.props.modalNextLink}
             </div>
           </Modal>
-        </React.Fragment>
+        </>
       )
     }
 
     return (
-      <div className={isHomepage ? `is-homepage` : ``}>
+      <>
         <SiteMetadata pathname={this.props.location.pathname} />
         <SkipNavLink css={styles.skipLink}>Skip to main content</SkipNavLink>
-        <Banner>
-          {/* !!! If you change the children of Banner remember to do the same in layout/layout-with-heading.js */}
-          <OutboundLink
-            href="https://www.gatsbyjs.com/behind-the-scenes/"
-            css={{ color: `#fff`, "&:hover": { color: `#fff` } }}
-          >
-            Watch now
-          </OutboundLink>
-          {`: “Behind the Scenes: What makes Gatsby Great”.`}
-        </Banner>
+        <Banner />
         <Navigation pathname={this.props.location.pathname} />
         <div
           className={`main-body`}
           css={{
-            paddingTop: presets.bannerHeight,
-            [presets.Md]: {
-              margin: `0 auto`,
-              paddingTop: isHomepage
-                ? presets.bannerHeight
-                : `calc(${presets.bannerHeight} + ${presets.headerHeight})`,
-            },
             paddingLeft: `env(safe-area-inset-left)`,
             paddingRight: `env(safe-area-inset-right)`,
+            paddingTop: presets.bannerHeight,
+            // make room for the mobile navigation
+            paddingBottom: `3.5rem`,
+            [presets.Md]: {
+              paddingTop: `calc(${presets.bannerHeight} + ${
+                presets.headerHeight
+              })`,
+              paddingBottom: 0,
+            },
           }}
         >
           <PageWithSidebar
@@ -184,7 +172,7 @@ class DefaultLayout extends React.Component {
           />
         </div>
         <MobileNavigation />
-      </div>
+      </>
     )
   }
 }
@@ -200,12 +188,12 @@ const styles = {
     overflow: `hidden`,
     position: `absolute`,
     zIndex: 100,
-    fontSize: `0.85rem`,
+    fontSize: presets.scale[1],
     ":focus": {
-      padding: `0.9rem`,
+      padding: rhythm(presets.space[4]),
       position: `fixed`,
-      top: 10,
-      left: 10,
+      top: rhythm(presets.space[6]),
+      left: rhythm(presets.space[6]),
       background: `white`,
       textDecoration: `none`,
       width: `auto`,