+++ www/src/views/showcase/featured-sites.js
@@ -8,8 +8,8 @@ import MdArrowForward from "react-icons/lib/md/arrow-forward"
 import ShowcaseItemCategories from "./showcase-item-categories"
 import FeaturedSitesIcon from "../../assets/featured-sites-icons.svg"
 import { ShowcaseIcon } from "../../assets/mobile-nav-icons"
-import { options, rhythm, scale } from "../../utils/typography"
-import presets, { colors } from "../../utils/presets"
+import { options, rhythm } from "../../utils/typography"
+import presets, { colors, space } from "../../utils/presets"
 import { svgStyles } from "../../utils/styles"
 import Button from "../../components/button"
 import ArrowForwardIcon from "react-icons/lib/md/arrow-forward"
@@ -28,7 +28,9 @@ class FeaturedSites extends Component {
       <section
         className="featured-sites"
         css={{
-          margin: `${rhythm(options.blockMarginBottom)} ${rhythm(3 / 4)} 0`,
+          margin: `${rhythm(options.blockMarginBottom)} ${rhythm(
+            presets.space[6]
+          )} 0`,
           position: `relative`,
           display: `none`,
           [presets.Lg]: {
@@ -65,8 +67,8 @@ class FeaturedSites extends Component {
           />
           <h1
             css={{
-              ...scale(1 / 5),
-              color: colors.gatsby,
+              fontSize: presets.scale[4],
+              color: colors.gray.dark,
               fontFamily: options.headerFontFamily.join(`,`),
               fontWeight: `bold`,
               marginRight: 30,
@@ -82,19 +84,14 @@ class FeaturedSites extends Component {
             css={{
               ...styles.withTitleHover,
               display: `none`,
+              fontSize: presets.scale[1],
               [presets.Sm]: {
                 display: `block`,
               },
               "&&": {
-                ...scale(-1 / 6),
-                boxShadow: `none`,
                 borderBottom: 0,
-                color: colors.lilac,
                 cursor: `pointer`,
-                fontFamily: options.headerFontFamily.join(`,`),
-                fontWeight: `normal`,
                 "&:hover": {
-                  background: `transparent`,
                   color: colors.gatsby,
                 },
               },
@@ -114,10 +111,9 @@ class FeaturedSites extends Component {
           >
             <div
               css={{
-                ...scale(-1 / 6),
                 color: colors.gray.calm,
+                fontSize: presets.scale[1],
                 marginRight: 15,
-                fontFamily: options.headerFontFamily.join(`,`),
                 display: `none`,
                 [presets.Md]: {
                   display: `block`,
@@ -148,8 +144,8 @@ class FeaturedSites extends Component {
               display: `flex`,
               overflowX: `scroll`,
               flexShrink: 0,
-              margin: `0 -${rhythm(3 / 4)}`,
-              padding: `3px ${rhythm(3 / 4)} 0`,
+              margin: `0 -${rhythm(space[6])}`,
+              padding: `3px ${rhythm(space[6])} 0`,
               ...styles.scrollbar,
             }}
           >
@@ -165,7 +161,10 @@ class FeaturedSites extends Component {
                   css={{
                     "&&": {
                       borderBottom: `none`,
-                      boxShadow: `none`,
+                      fontSize: presets.scale[3],
+                      fontWeight: `bold`,
+                      color: colors.gray.dark,
+                      fontFamily: options.headerFontFamily.join(`,`),
                       transition: `box-shadow .3s cubic-bezier(.4,0,.2,1), transform .3s cubic-bezier(.4,0,.2,1)`,
                       "&:hover": { ...styles.screenshotHover },
                     },
@@ -189,7 +188,7 @@ class FeaturedSites extends Component {
                 </Link>
                 <div
                   css={{
-                    ...scale(-1 / 6),
+                    fontSize: presets.scale[1],
                     color: colors.gray.calm,
                     fontWeight: `normal`,
                     [presets.Lg]: {
@@ -221,9 +220,9 @@ class FeaturedSites extends Component {
               <a
                 href="#showcase"
                 css={{
-                  marginRight: `${rhythm(3 / 4)} !important`,
+                  marginRight: `${rhythm(space[6])} !important`,
                   backgroundColor: hex2rgba(colors.ui.light, 0.25),
-                  borderRadius: presets.radius,
+                  borderRadius: presets.radii[1],
                   textAlign: `center`,
                   "&&": {
                     border: `1px solid ${colors.ui.light}`,
@@ -243,7 +242,7 @@ class FeaturedSites extends Component {
               >
                 <div
                   css={{
-                    borderRadius: presets.radius,
+                    borderRadius: presets.radii[1],
                     display: `flex`,
                     alignItems: `center`,
                     position: `relative`,
@@ -290,7 +289,7 @@ class FeaturedSites extends Component {
               position: `absolute`,
               top: `0`,
               bottom: rhythm(options.blockMarginBottom),
-              right: `-${rhythm(3 / 4)}`,
+              right: `-${rhythm(space[6])}`,
               width: 60,
               pointerEvents: `none`,
               background: `linear-gradient(90deg, rgba(0,0,0,0) 0%, rgba(255,255,255,1) 100%)`,