+++ packages/gatsby-dev-cli/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.11](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-dev-cli/compare/gatsby-dev-cli@2.4.10...gatsby-dev-cli@2.4.11) (2019-03-11)
+
+**Note:** Version bump only for package gatsby-dev-cli
+
 ## [2.4.10](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-dev-cli/compare/gatsby-dev-cli@2.4.9...gatsby-dev-cli@2.4.10) (2019-02-19)
 
 **Note:** Version bump only for package gatsby-dev-cli