+++ packages/gatsby-remark-prismjs/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.5](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-remark-prismjs/compare/gatsby-remark-prismjs@3.2.4...gatsby-remark-prismjs@3.2.5) (2019-03-11)
+
+**Note:** Version bump only for package gatsby-remark-prismjs
+
 ## [3.2.4](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-remark-prismjs/compare/gatsby-remark-prismjs@3.2.3...gatsby-remark-prismjs@3.2.4) (2019-02-01)
 
 **Note:** Version bump only for package gatsby-remark-prismjs